upstream merge: all: java: remove useless type boxing (88 files)
1 minor commit (no functionality changes), manually adjusted and
reviewed:
"Improved: Remove all unnecessary boxing and unboxing in Java classes
(OFBIZ-10504)"
SCIPIO: Several manual cleanups on top. This removes maintenance noise
and awful code.

Upstream-Commit: 56419cdfa8b3a796b85a6a38bb30f98dedd08c39
Upstream-Repo: https://github.com/apache/ofbiz-framework.git
Upstream-Branch: trunk

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/556

+++ applications/content/src/org/ofbiz/content/ContentManagementServices.java
@@ -880,9 +880,9 @@ public class ContentManagementServices {
         String contentIdTo = (String)context.get("contentIdTo");
         Integer seqInc = (Integer)context.get("seqInc");
         if (seqInc == null) {
-            seqInc = Integer.valueOf(100);
+            seqInc = 100;
         }
-        int seqIncrement = seqInc.intValue();
+        int seqIncrement = seqInc;
         List<String> typeList = UtilGenerics.checkList(context.get("typeList"));
         if (typeList == null) {
             typeList = new LinkedList<String>();
@@ -917,7 +917,7 @@ public class ContentManagementServices {
                                 try {
                                     GenericValue prevValue = listAll.get(i-1);
                                     Long prevSeqNum = (Long)prevValue.get("sequenceNum");
-                                    prevValue.put("sequenceNum", Long.valueOf(seqNum));
+                                    prevValue.put("sequenceNum", (long) seqNum);
                                     prevValue.store();
                                     contentAssoc.put("sequenceNum", prevSeqNum);
                                     contentAssoc.store();
@@ -929,20 +929,20 @@ public class ContentManagementServices {
                             if (i < listAll.size()) {
                                 // Swap with next entry
                                 GenericValue nextValue = listAll.get(i+1);
-                                nextValue.put("sequenceNum", Long.valueOf(seqNum));
+                                nextValue.put("sequenceNum", (long) seqNum);
                                 nextValue.store();
                                 seqNum += seqIncrement;
-                                contentAssoc.put("sequenceNum", Long.valueOf(seqNum));
+                                contentAssoc.put("sequenceNum", (long) seqNum);
                                 contentAssoc.store();
                                 i++; // skip next one
                             }
                         }
                     } else {
-                        contentAssoc.put("sequenceNum", Long.valueOf(seqNum));
+                        contentAssoc.put("sequenceNum", (long) seqNum);
                         contentAssoc.store();
                     }
                 } else {
-                    contentAssoc.put("sequenceNum", Long.valueOf(seqNum));
+                    contentAssoc.put("sequenceNum", (long) seqNum);
                     contentAssoc.store();
                 }
                 seqNum += seqIncrement;
@@ -1002,7 +1002,7 @@ public class ContentManagementServices {
                 serviceIn.put("userLogin", userLogin);
                 serviceIn.put("contentIdTo", contentId);
                 serviceIn.put("contentAssocTypeId", "SUB_CONTENT");
-                serviceIn.put("sequenceNum", Long.valueOf(50));
+                serviceIn.put("sequenceNum", 50L);
                 try {
                     dispatcher.runSync("persistContentAndAssoc", serviceIn);
                 } catch (ServiceAuthException e) {
@@ -1030,7 +1030,7 @@ public class ContentManagementServices {
         String startContentId = (String)context.get("contentId");
         try {
             int leafCount = ContentManagementWorker.updateStatsTopDown(delegator, startContentId, typeList);
-            result.put("leafCount", Integer.valueOf(leafCount));
+            result.put("leafCount", leafCount);
         } catch (GenericEntityException e) {
             Debug.logError(e, module);
             return ServiceUtil.returnError(e.toString());
@@ -1337,11 +1337,11 @@ public class ContentManagementServices {
         }
         Long leafCount = (Long)content.get("childLeafCount");
         if (leafCount == null) {
-            content.set("childLeafCount", Long.valueOf(0));
+            content.set("childLeafCount", 0L);
         }
         Long branchCount = (Long)content.get("childBranchCount");
         if (branchCount == null) {
-            content.set("childBranchCount", Long.valueOf(0));
+            content.set("childBranchCount", 0L);
         }
 
         //content.store();
@@ -1363,7 +1363,7 @@ public class ContentManagementServices {
             }
             Long leafCount = (Long)content.get("childLeafCount");
             if (leafCount == null) {
-                leafCount = Long.valueOf(0);
+                leafCount = 0L;
             }
             int changeLeafCount = leafCount.intValue() + 1;
             int changeBranchCount = 1;
@@ -1389,7 +1389,7 @@ public class ContentManagementServices {
             }
             Long leafCount = (Long)content.get("childLeafCount");
             if (leafCount == null) {
-                leafCount = Long.valueOf(0);
+                leafCount = 0L;
             }
             int changeLeafCount = -1 * leafCount.intValue() - 1;
             int changeBranchCount = -1;
@@ -1482,7 +1482,7 @@ public class ContentManagementServices {
         } else {
             Debug.logWarning("Don't know anything about useTimeUomId [" + useTimeUomId + "], defaulting to month", module);
         }
-        calendar.add(field, useTime.intValue());
+        calendar.add(field, useTime);
         thruDate = new Timestamp(calendar.getTimeInMillis());
         contentRole.set("thruDate", thruDate);
         try {
@@ -1510,7 +1510,7 @@ public class ContentManagementServices {
         String productId = (String) context.get("productId");
         Integer qty = (Integer) context.get("quantity");
         if (qty == null) {
-            qty = Integer.valueOf(1);
+            qty = 1;
         }
 
         Timestamp orderCreatedDate = (Timestamp) context.get("orderCreatedDate");
@@ -1542,7 +1542,7 @@ public class ContentManagementServices {
         Long useTime = (Long) productContent.get("useTime");
         Integer newUseTime = null;
         if (UtilValidate.isNotEmpty(useTime)) {
-            newUseTime = Integer.valueOf(useTime.intValue() * qty.intValue());
+            newUseTime = useTime.intValue() * qty;
         }
         context.put("useTime", newUseTime);
         context.put("useTimeUomId", productContent.get("useTimeUomId"));
@@ -1593,7 +1593,7 @@ public class ContentManagementServices {
                         .filterByDate().queryCount();
                 if (productContentCount > 0) {
                     context.put("productId", productId);
-                    context.put("quantity", Integer.valueOf(qty.intValue()));
+                    context.put("quantity", qty.intValue());
                     Map<String, Object> ctx = subscriptionModel.makeValid(context, ModelService.IN_PARAM);
                     dispatcher.runSync("updateContentSubscriptionByProduct", ctx);
                 }
