+++ applications/content/src/org/ofbiz/content/compdoc/CompDocEvents.java
@@ -134,7 +134,7 @@ public class CompDocEvents {
     }
 
     public static String padNumberWithLeadingZeros(Long num, Integer padLen) {
-        String s = UtilFormatOut.formatPaddedNumber(num.longValue(), padLen.intValue());
+        String s = UtilFormatOut.formatPaddedNumber(num, padLen);
         return s;
     }
