+++ applications/content/src/org/ofbiz/content/content/ContentSearchEvents.java
@@ -55,17 +55,17 @@ public class ContentSearchEvents {
         ContentSearchOptions contentSearchOptions = ContentSearchSession.getContentSearchOptions(session);
 
         Integer viewIndexInteger = contentSearchOptions.getViewIndex();
-        if (viewIndexInteger != null) viewIndex = viewIndexInteger.intValue();
+        if (viewIndexInteger != null) viewIndex = viewIndexInteger;
         Integer viewSizeInteger = contentSearchOptions.getViewSize();
-        if (viewSizeInteger != null) viewSize = viewSizeInteger.intValue();
+        if (viewSizeInteger != null) viewSize = viewSizeInteger;
 
         lowIndex = viewIndex * viewSize;
         highIndex = (viewIndex + 1) * viewSize;
 
         // setup resultOffset and maxResults, noting that resultOffset is 1 based, not zero based as these numbers
-        Integer resultOffset = Integer.valueOf(lowIndex + 1);
+        Integer resultOffset = lowIndex + 1;
         // SCIPIO: FIXME: maxResults of viewSize ensures pagination is disabled. The whole thing needs rewriting to support pagination.
-        Integer maxResults = Integer.valueOf(viewSize);
+        Integer maxResults = viewSize;
 
         // ========== Do the actual search
         ArrayList<String> contentIds = null;
@@ -89,7 +89,7 @@ public class ContentSearchEvents {
 
             Integer totalResults = contentSearchContext.getTotalResults();
             if (totalResults != null) {
-                listSize = totalResults.intValue();
+                listSize = totalResults;
             }
         }
 
@@ -105,11 +105,11 @@ public class ContentSearchEvents {
         Map<String, Object> result = new HashMap<String, Object>();
 
         result.put("contentIds", contentIds);
-        result.put("viewIndex", Integer.valueOf(viewIndex));
-        result.put("viewSize", Integer.valueOf(viewSize));
-        result.put("listSize", Integer.valueOf(listSize));
-        result.put("lowIndex", Integer.valueOf(lowIndex));
-        result.put("highIndex", Integer.valueOf(highIndex));
+        result.put("viewIndex", viewIndex);
+        result.put("viewSize", viewSize);
+        result.put("listSize", listSize);
+        result.put("lowIndex", lowIndex);
+        result.put("highIndex", highIndex);
         result.put("searchConstraintStrings", searchConstraintStrings);
         result.put("searchSortOrderString", searchSortOrderString);
